Fixed multiple issues in with separate installation script. Base setup now contains git, virt-manager and firefox Chrome installer is disabled per default as removal of pulseaudio required to make it work is currently no option

+++ scripts/install/installChrome.sh
@@ -1,9 +1,13 @@
-#!/usr/bin/env bash
+#!/bin/bash
 
-sudo apt update && sudo apt upgrade
-sudo apt install mesa-utils fonts-liberation
-sudo apt purge pulseaudio libpulse0 && sudo apt autoremove
+echo -e "\n\nInstallation is currently disabled as removal of pulseaudio cripples others software packages.\nUse \"-f\" to force the installation.\n"
 
-wget https://dl.google.com/linux/direct/google-chrome-stable_current_amd64.deb -O /tmp/google-chrome-stable_current_amd64.deb
-sudo dpkg -i /tmp/google-chrome-stable_current_amd64.deb
-rm /tmp/google-chrome-stable_current_amd64.deb
+if [[ "$1" == "force" ]]; then
+    sudo apt update && sudo apt upgrade
+    sudo apt install mesa-utils fonts-liberation
+    sudo apt purge pulseaudio libpulse0
+
+    wget https://dl.google.com/linux/direct/google-chrome-stable_current_amd64.deb -O /tmp/google-chrome-stable_current_amd64.deb
+    sudo dpkg -i /tmp/google-chrome-stable_current_amd64.deb
+    rm /tmp/google-chrome-stable_current_amd64.deb
+fi
